backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
gunicorn
dj-database-url
python-dotenv
supabase==1.0.0
django-cors-headers
